firebase model structure
